requirements.txt
selenium
pyyaml
python-dotenv
colorama
pytest
cryptography
pyinstaller
webdriver-manager
